# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba Técnica: Sistema de Búsqueda de Vuelos",
        "description": "API para gestionar reservas de vuelos y obtener información de aeropuertos y vuelos.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/airports": {
            "post": {
                "tags": [
                    "Aeropuertos"
                ],
                "summary": "Obtener aeropuertos",
                "operationId": "30dc94b28459ecbf722b181220d42108",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "medellin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aeropuertos recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Airport"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El parámetro code es requerido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/flights": {
            "post": {
                "tags": [
                    "Vuelos"
                ],
                "summary": "Obtener vuelos",
                "operationId": "64a6f7e7f585edcf896b3c96bc72c7ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "itinerary",
                                    "qtyPassengers",
                                    "adult",
                                    "searchs"
                                ],
                                "properties": {
                                    "qtyPassengers": {
                                        "description": "Número total de pasajeros",
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "adult": {
                                        "description": "Número de adultos",
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "child": {
                                        "description": "Número de niños",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "baby": {
                                        "description": "Número de bebés",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "searchs": {
                                        "description": "Número de búsquedas a realizar",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "itinerary": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "departureCity": {
                                                    "description": "Ciudad de salida",
                                                    "type": "string",
                                                    "example": "MDE"
                                                },
                                                "arrivalCity": {
                                                    "description": "Ciudad de llegada",
                                                    "type": "string",
                                                    "example": "BOG"
                                                },
                                                "hour": {
                                                    "description": "Hora de salida",
                                                    "type": "string",
                                                    "example": "2024-10-25T10:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vuelos recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Flight"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El parámetro itinerary es requerido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Airport": {
                "title": "Airport",
                "description": "Modelo de aeropuerto",
                "properties": {
                    "ciudad": {
                        "description": "Ciudad del aeropuerto",
                        "type": "string",
                        "example": "Medellin"
                    },
                    "nombre": {
                        "description": "Nombre del aeropuerto",
                        "type": "string",
                        "example": "Jose Maria Cordova"
                    },
                    "país": {
                        "description": "Nombre del país",
                        "type": "string",
                        "example": "Colombia"
                    },
                    "codigo IATA": {
                        "description": "Codigo IATA de la ciudad",
                        "type": "string",
                        "example": "MDE"
                    }
                },
                "type": "object"
            },
            "Flight": {
                "title": "Flight",
                "description": "Modelo de vuelo",
                "properties": {
                    "dateOfDeparture": {
                        "description": "Fecha de salida",
                        "type": "string",
                        "example": "2024-11-20"
                    },
                    "timeOfDeparture": {
                        "description": "Hora de salida",
                        "type": "string",
                        "example": "10:30"
                    },
                    "dateOfArrival": {
                        "description": "Fecha de llegada",
                        "type": "string",
                        "example": "2024-11-20"
                    },
                    "timeOfArrival": {
                        "description": "Hora de llegada",
                        "type": "string",
                        "example": "14:45"
                    },
                    "marketingCarrier": {
                        "description": "Código de la aerolínea",
                        "type": "string",
                        "example": "AA"
                    },
                    "flightOrtrainNumber": {
                        "description": "Número de vuelo",
                        "type": "string",
                        "example": "1234"
                    },
                    "locationId": {
                        "properties": {
                            "departureCity": {
                                "description": "Ciudad de salida",
                                "type": "string",
                                "example": "BOG"
                            },
                            "arrivalCity": {
                                "description": "Ciudad de llegada",
                                "type": "string",
                                "example": "MIA"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}